framework-comparison/requirements.txt
strands-agents>=1.25.0
langgraph>=0.4.0
langchain-aws>=0.2.0
langchain-core>=0.3.0
